+++ tests/Api/ApiTest.php
@@ -154,6 +154,7 @@ class ApiTest extends TestCase
             [Entities\ScheduledPayment::class],
             [Entities\SubscriptionReactivation::class, null],
             [Entities\Subscription::class],
+            [Entities\SubscriptionCancel::class, null],
             [Entities\SubscriptionCancellation::class, null],
             [Entities\Transaction::class],
             [Entities\Website::class],